Move export button to FundView and update export service
